start making an autoloading plugin out of this
