+++ classes/class-user-account-management-transient.php
@@ -1,0 +1,99 @@
+<?php
+/**
+ * Class file for the User_Account_Management_Transient class.
+ *
+ * @file
+ */
+
+if ( ! class_exists( 'User_Account_Management' ) ) {
+	die();
+}
+
+/**
+ * Class to store all theme/plugin transients as an array in one WordPress transient
+ **/
+class User_Account_Management_Transient {
+
+	protected $name;
+	public $cache_expiration;
+
+	/**
+	 * Constructor which sets cache options and the name of the field that lists this plugin's cache keys.
+	 *
+	 * @param string $name The name of the field that lists all cache keys.
+	 */
+	public function __construct( $name, $cache_expiration = 2592000 ) {
+		$this->name             = $name;
+		$this->cache_expiration = $cache_expiration; // cache expiration in seconds
+		$this->cache_prefix     = esc_sql( 'acct_mgmt_' );
+	}
+
+	/**
+	 * Get the transient that lists all the other transients for this plugin.
+	 *
+	 * @return mixed value of transient. False of empty, otherwise array.
+	 */
+	public function all_keys() {
+		return get_transient( $this->name );
+	}
+
+	/**
+	 * Set individual transient, and add its key to the list of this plugin's transients.
+	 *
+	 * @param string $cachekey the key for this cache item
+	 * @param mixed $value the value of the cache item
+	 * @param int $cache_expiration. How long the plugin key cache, and this individual item cache, should last before expiring.
+	 * @return mixed value of transient. False of empty, otherwise array.
+	 */
+	public function set( $cachekey, $value ) {
+
+		$prefix   = $this->cache_prefix;
+		$cachekey = $prefix . $cachekey;
+
+		$keys   = $this->all_keys();
+		$keys[] = $cachekey;
+		set_transient( $this->name, $keys, $this->cache_expiration );
+
+		return set_transient( $cachekey, $value, $this->cache_expiration );
+	}
+
+	/**
+	 * Get the individual cache value
+	 *
+	 * @param string $cachekey the key for this cache item
+	 * @return mixed value of transient. False of empty, otherwise array.
+	 */
+	public function get( $cachekey ) {
+		$prefix   = $this->cache_prefix;
+		$cachekey = $prefix . $cachekey;
+		return get_transient( $cachekey );
+	}
+
+	/**
+	 * Delete the individual cache value
+	 *
+	 * @param string $cachekey the key for this cache item
+	 * @return bool True if successful, false otherwise.
+	 */
+	public function delete( $cachekey ) {
+		$prefix   = $this->cache_prefix;
+		$cachekey = $prefix . $cachekey;
+		return delete_transient( $cachekey );
+	}
+
+	/**
+	 * Delete the entire cache for this plugin
+	 *
+	 * @return bool True if successful, false otherwise.
+	 */
+	public function flush() {
+		$keys   = $this->all_keys();
+		$result = true;
+		foreach ( $keys as $key ) {
+			$result = delete_transient( $key );
+		}
+		$result = delete_transient( $this->name );
+		return $result;
+	}
+
+}
